% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means-clustering.R
\name{k.means.clustering}
\alias{k.means.clustering}
\title{Example on how to use k-means clustering}
\usage{
k.means.clustering(arg)
}
\arguments{
\item{arg}{; A numeric vector to calculate the arithmetic mean over.}
}
\value{
numeric ; The Arithmetic mean of 'src.vector'.
}
\description{
Example on how to use k-means clustering
}
